Update Client layer APIs to validate context (#28)
Update the `Sntp_SendTimeRequest` and `Sntp_ReceiveTimeResponse` APIs to validate the members of the `SntpContext_t` parameter passed toas input to avoid NULL dereferencing issues for pointer members of the context as well as fail immediately for an uininitialized/invalid context instead of attempting to perform the send/recv operation. This PR adds an `SntpErrorContextNotInitialized` error code to handle the case when either an invalid or unininitialized context is passed to the APIs.